fix: circular dependency on select prisma
