+++ apps/bilan-carbone/src/db/organization.ts
@@ -6,6 +6,7 @@ import type { Organization, OrganizationVersion, Site } from '@repo/db-common'
 import { Prisma } from '@repo/db-common'
 import { Environment, EstablishmentType, UserStatus } from '@repo/db-common/enums'
 import { prismaClient } from './client.server'
+import { OrganizationVersionWithOrganizationSelect } from './organization.select'
 import { deleteStudy } from './study'
 
 export type OrganizationVersionWithOrganization = OrganizationVersion & {
@@ -18,82 +19,6 @@ export type OrganizationVersionWithOrganization = OrganizationVersion & {
   }
 }
 export type OrganizationVersionWithOrganizationWithoutSites = OrganizationVersion & { organization: Organization }
-
-export const OrganizationVersionWithOrganizationSelect = {
-  id: true,
-  createdAt: true,
-  updatedAt: true,
-  organizationId: true,
-  isCR: true,
-  activatedLicence: true,
-  onboarded: true,
-  onboarderId: true,
-  environment: true,
-  parentId: true,
-  parent: {
-    select: {
-      activatedLicence: true,
-    },
-  },
-  organization: {
-    select: {
-      oldBCId: true,
-      id: true,
-      name: true,
-      createdAt: true,
-      updatedAt: true,
-      importedFileDate: true,
-      wordpressId: true,
-      sites: {
-        select: {
-          name: true,
-          etp: true,
-          ca: true,
-          id: true,
-          createdAt: true,
-          updatedAt: true,
-          organizationId: true,
-          oldBCId: true,
-          postalCode: true,
-          city: true,
-          volunteerNumber: true,
-          beneficiaryNumber: true,
-          establishmentId: true,
-          establishmentYear: true,
-          studentNumber: true,
-          academy: true,
-          establishmentType: true,
-          superficy: true,
-          address: true,
-          cncId: true,
-          country: true,
-          cnc: {
-            select: {
-              cncCode: true,
-              seances: true,
-              entrees2024: true,
-              entrees2023: true,
-              semainesActivite: true,
-              latitude: true,
-              longitude: true,
-              cncVersionId: true,
-            },
-          },
-        },
-        orderBy: { createdAt: Prisma.SortOrder.asc },
-      },
-    },
-  },
-  userAccounts: {
-    select: {
-      user: {
-        select: {
-          level: true,
-        },
-      },
-    },
-  },
-}
 
 export const getOrgNameByOrgVersionId = async (id: string | null) => {
   if (!id) {
